SpeechToText/requirements.txt
openai
python-dotenv
sounddevice
scipy
keyboard
fastapi
uvicorn[standard]
python-multipart
